Remove teleport delay, and fix issue of players sometimes teleporting to wrong area.
